Extract API provider storage
