+++ tests/test-custom-api.js
@@ -37,27 +37,28 @@ await import('../js/provider-panels.js');
 // ─── 1. api.js source inspection ───
 console.log('1. api.js source inspection');
 const apiSrc = read('js/api.js');
-assert('getCustomApiUrl exists', apiSrc.includes('function getCustomApiUrl()'));
-assert('setCustomApiUrl exists', apiSrc.includes('function setCustomApiUrl('));
-assert('getCustomApiKey exists', apiSrc.includes('function getCustomApiKey()'));
-assert('saveCustomApiKey exists', apiSrc.includes('function saveCustomApiKey('));
-assert('hasCustomApiKey exists', apiSrc.includes('function hasCustomApiKey()'));
-assert('getCustomApiModel exists', apiSrc.includes('function getCustomApiModel()'));
-assert('setCustomApiModel exists', apiSrc.includes('function setCustomApiModel('));
-assert('getCustomApiModelDisplay exists', apiSrc.includes('function getCustomApiModelDisplay()'));
+const apiProviderStorageSrc = read('js/api-provider-storage.js');
+assert('getCustomApiUrl exists', apiProviderStorageSrc.includes('function getCustomApiUrl()'));
+assert('setCustomApiUrl exists', apiProviderStorageSrc.includes('function setCustomApiUrl('));
+assert('getCustomApiKey exists', apiProviderStorageSrc.includes('function getCustomApiKey()'));
+assert('saveCustomApiKey exists', apiProviderStorageSrc.includes('function saveCustomApiKey('));
+assert('hasCustomApiKey exists', apiProviderStorageSrc.includes('function hasCustomApiKey()'));
+assert('getCustomApiModel exists', apiProviderStorageSrc.includes('function getCustomApiModel()'));
+assert('setCustomApiModel exists', apiProviderStorageSrc.includes('function setCustomApiModel('));
+assert('getCustomApiModelDisplay exists', apiProviderStorageSrc.includes('function getCustomApiModelDisplay()'));
 assert('fetchCustomApiModels exists', apiSrc.includes('function fetchCustomApiModels('));
 assert('validateCustomApiKey exists', apiSrc.includes('function validateCustomApiKey('));
 assert('callCustomAPI exists', apiSrc.includes('function callCustomAPI('));
-assert('hasAIProvider handles custom', apiSrc.includes("provider === 'custom') return hasCustomApiKey()"));
-assert('hasAIProvider custom requires URL', apiSrc.includes("hasCustomApiKey() && !!getCustomApiUrl()"));
+assert('hasAIProvider handles custom', apiProviderStorageSrc.includes("provider === 'custom') return hasCustomApiKey()"));
+assert('hasAIProvider custom requires URL', apiProviderStorageSrc.includes("hasCustomApiKey() && !!getCustomApiUrl()"));
 assert('getActiveModelId handles custom', apiSrc.includes("provider === 'custom') return getCustomApiModel()"));
 assert('getActiveModelDisplay handles custom', apiSrc.includes("provider === 'custom') return getCustomApiModelDisplay()"));
 assert('callClaudeAPI handles custom', apiSrc.includes("provider === 'custom') return callCustomAPI("));
 assert('supportsWebSearch false for custom', apiSrc.includes("provider === 'custom') return false"));
 assert('supportsVision true for custom', apiSrc.includes("provider === 'custom') return true"));
 assert('callCustomAPI routes through proxy', apiSrc.includes("'Custom', opts,\n    {}"));
-assert('saveCustomApiKey uses encryptedSetItem', apiSrc.includes("encryptedSetItem('labcharts-custom-key'"));
-assert('getCustomApiKey uses getCachedKey', apiSrc.includes("getCachedKey('labcharts-custom-key')"));
+assert('saveCustomApiKey uses encryptedSetItem', apiProviderStorageSrc.includes("encryptedSetItem('labcharts-custom-key'"));
+assert('getCustomApiKey uses getCachedKey', apiProviderStorageSrc.includes("getCachedKey('labcharts-custom-key')"));
 
 // ─── 2. Window function exports ───
 console.log('\n2. Window function exports');
